refactor: type transcript and store events
