+++ data/recordingsStore.ts
@@ -2,6 +2,7 @@ import { Platform } from 'react-native';
 import { AudioFile } from '@/types/audio';
 import { Tag } from '@/types/tag';
 import { Folder } from '@/types/folder';
+import { StoreEvent } from '@/types/store';
 import { AudioStorageService } from '@/services/audioStorage';
 import { FolderService } from '@/services/folderService';
 import { StorageService } from '@/services/storageService';
@@ -11,7 +12,7 @@ const TAGS_KEY = 'rg.tags.v1';
 
 // Store change listeners
 type StoreChangeListener = () => void;
-type StoreChangeListenerWithEvent = (event?: any) => void;
+type StoreChangeListenerWithEvent = (event?: StoreEvent) => void;
 const storeChangeListeners: StoreChangeListener[] = [];
 const storeChangeListenersWithEvent: StoreChangeListenerWithEvent[] = [];
 
@@ -42,7 +43,7 @@ export class RecordingsStore {
     return Date.now().toString() + '_' + Math.random().toString(36).substring(2, 15);
   }
 
-  static notifyStoreChanged(fromBroadcast: boolean = false, event?: any): void {
+  static notifyStoreChanged(fromBroadcast: boolean = false, event?: StoreEvent): void {
     // Broadcast to other tabs/windows on web platform
     if (Platform.OS === 'web' && typeof window !== 'undefined' && typeof window.BroadcastChannel !== 'undefined' && !fromBroadcast && syncChannel && event) {
       try {
@@ -100,7 +101,7 @@ export class RecordingsStore {
 
   static async moveRecordingToFolder(recordingId: string, folderId: string | null): Promise<AudioFile | null> {
     try {
-      const updatedRecording = await AudioStorageService.updateAudioFile(recordingId, { folderId });
+      const updatedRecording = await AudioStorageService.updateAudioFile(recordingId, { folderId: folderId ?? undefined });
       if (!updatedRecording) {
         throw new Error('Recording not found');
       }
@@ -320,7 +321,7 @@ export class RecordingsStore {
       // Update each recording individually
       for (const recording of updatedRecordings) {
         if (recording.folderId !== recordings.find(r => r.id === recording.id)?.folderId) {
-          await AudioStorageService.updateAudioFile(recording.id, { folderId: recording.folderId });
+          await AudioStorageService.updateAudioFile(recording.id, { folderId: recording.folderId ?? undefined });
         }
       }
