+++ services/transcriptService.ts
@@ -1,4 +1,4 @@
-import { Transcript, TranscriptSegment } from '@/types/transcript';
+import { Transcript, TranscriptSegment, TranscriptSegmentPayload } from '@/types/transcript';
 import { StorageService } from './storageService';
 import logger from '@/utils/logger';
 
@@ -31,7 +31,7 @@ export class TranscriptService {
     }
   }
 
-  static async storeTranscriptWithVerification(transcriptId: string, data: any): Promise<boolean> {
+  static async storeTranscriptWithVerification(transcriptId: string, data: Transcript): Promise<boolean> {
     const maxRetries = 3;
     const retryDelay = 200;
     
@@ -125,7 +125,7 @@ export class TranscriptService {
     return false;
   }
 
-  static async storeTranscript(transcriptId: string, data: any): Promise<boolean> {
+  static async storeTranscript(transcriptId: string, data: Transcript): Promise<boolean> {
     try {
       return await this.storeTranscriptWithVerification(transcriptId, data);
     } catch (error) {
@@ -200,7 +200,7 @@ export class TranscriptService {
       if (!transcript?.segments || !Array.isArray(transcript.segments)) {
         logger.log('⚠️ Retrieved transcript segments are null or missing.');
       } else {
-        const validSegments = transcript.segments.filter(s => s && typeof s.text === 'string' && s.text.trim().length > 0);
+        const validSegments = transcript.segments.filter((s: TranscriptSegment) => s && typeof s.text === 'string' && s.text.trim().length > 0);
         logger.log('✅ Valid segments count:', validSegments.length);
         if (validSegments.length !== transcript.segments.length) {
           logger.log('⚠️ Some segments have invalid text');
@@ -254,7 +254,7 @@ export class TranscriptService {
 
   static async createTranscriptFromWhisper(
     fileId: string,
-    segments: any[],
+    segments: TranscriptSegmentPayload[],
     fullText: string,
     language: string,
     duration: number
@@ -285,7 +285,7 @@ export class TranscriptService {
       if (segments && Array.isArray(segments)) {
         logger.log('🔍 Segments validation passed, length:', segments.length);
         
-        processedSegments = segments.map((segment, index) => {
+        processedSegments = segments.map((segment: TranscriptSegmentPayload, index) => {
           logger.log(`🔍 Processing segment ${index}:`, {
             segment,
             hasText: segment && typeof segment.text === 'string',
