+++ types/store.ts
@@ -1,0 +1,4 @@
+export interface StoreEvent<T extends string = string, P = unknown> {
+  type: T;
+  payload?: P;
+}
